+++ package.json
@@ -46,9 +46,9 @@
     "@babel/preset-env": "^7.3.1",
     "@babel/preset-react": "^7.0.0",
     "@playwright/test": "^1.24.1",
-    "@rollup/plugin-commonjs": "^23.0.0",
-    "@rollup/plugin-node-resolve": "^15.0.0",
-    "@rollup/plugin-replace": "^5.0.0",
+    "@rollup/plugin-node-resolve": "^15.0.1",
+    "@rollup/plugin-replace": "^5.0.1",
+    "@rollup/plugin-terser": "^0.2.0",
     "@storybook/addon-backgrounds": "^6.5.12",
     "@storybook/addon-controls": "^6.5.12",
     "@storybook/addon-viewport": "^6.5.12",
@@ -89,15 +89,14 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-syntax-highlighter": "^15.5.0",
-    "rollup": "^2.72.1",
+    "rollup": "^3.7.0",
     "rollup-plugin-copy": "^3.4.0",
-    "rollup-plugin-scss": "^3.0.0",
-    "rollup-plugin-terser": "^7.0.2",
+    "rollup-plugin-scss": "^4.0.0",
     "sass": "^1.52.3",
     "sass-loader": "^13.0.0",
     "svgo": "^2.8.0",
     "testcafe": "1.19.0",
-    "vitest": "^0.16.0",
+    "vitest": "^0.25.7",
     "wait-on": "^6.0.1",
     "webpack": "^5.72.1"
   },
