+++ src/components/10-atoms/button/index.react.js
@@ -1,5 +1,5 @@
 import withReact from '../../../utils/with-react';
-import AXAButton from './index';
+import AXAButton from './index.wc';
 
 export default (createElement, version) =>
   withReact(createElement, AXAButton, version);
