+++ src/components/10-atoms/button/package.json
@@ -5,13 +5,17 @@
   "author": "Pattern Warriors",
   "homepage": "https://github.com/axa-ch-webhub-cloud/pattern-library/tree/develop/src/components/10-atoms/button#readme",
   "license": "Copyright 2019 AXA Versicherungen AG",
+  "type": "module",
+  "main": "lib/index.js",
+  "exports": {
+    ".": "./lib/index.js"
+  },
   "directories": {
     "lib": "lib"
   },
   "scripts": {
     "prepublishOnly": "../../../../scripts/build/prepublish.js",
-    "build": "rollup --silent --config ../../../../rollup.config.js",
-    "build-dist": "rollup --silent --config ../../../../rollup.config.dist.js",
+    "build": "rollup --silent --config ../../../../rollup.config.js --bundleConfigAsCjs",
     "postpublish": "../../../../scripts/build/postpublish.js",
     "watch": "rollup --watch --silent --config ../../../../rollup.config.js",
     "genPackagelock": "npm i --package-lock-only"
